Fixed autocomplete styling / documentation
Using the suggested classes for an autocomplete style typeahead
resulted in the component being entirely relatively or absolutely
positioned, meaning the options took up space on the page and pushed
content below or the input element itself was lifted from the page
meaning the content below slotted underneath it.

To fix this:
* Taken the input element out of the typeahead list.
* Removed the typeahead input class to prevent the overriding of the
default form control styles on the input element.